sparkle::core::error

Enum Error

pub enum Error {

    Timeout {
        message: String,
        timeout_ms: u64,
    },
    ElementNotFound {
        selector: String,
    },
    StrictModeViolation {
        selector: String,
        count: usize,
    },
    WebDriver(WebDriverError),
    Network(String),
    Navigation(String),
    Serialization(Error),
    InvalidArgument(String),
    BrowserClosed,
    ContextClosed,
    PageClosed,
    FrameDetached,
    ElementNotAttached,
    ElementNotVisible {
        selector: String,
    },
    ElementNotEnabled {
        selector: String,
    },
    ElementNotEditable {
        selector: String,
    },
    JsEvaluation(String),
    File(Error),
    InvalidUrl(ParseError),
    DownloadFailed(String),
    ScreenshotFailed(String),
    VideoRecording(String),
    Tracing(String),
    UnsupportedBrowser(String),
    BrowserNotFound(String),
    NotImplemented(String),
    ActionFailed(String),
    Internal(String),

}

The main error type for Sparkle operations

This enum represents all possible errors that can occur while using Sparkle. It follows Playwrightโs error hierarchy while being idiomatic to Rust.

Variants

Timeout

Timeout occurred while waiting for an operation to complete

Fields

message: String

Description of what operation timed out

timeout_ms: u64

Timeout duration in milliseconds

ElementNotFound

Element could not be found using the given selector

Fields

selector: String

The selector that was used

StrictModeViolation

Multiple elements found when expecting a single element (strict mode)

Fields

selector: String

The selector that matched multiple elements

count: usize

Number of elements found

WebDriver(WebDriverError)

Error from the underlying WebDriver implementation

Network(String)

Network-related error

Navigation(String)

Navigation failed

Serialization(Error)

JSON serialization/deserialization error

InvalidArgument(String)

Invalid argument provided to a function

BrowserClosed

Browser instance has been closed

ContextClosed

Browser context has been closed

PageClosed

Page has been closed

FrameDetached

Frame has been detached from the page

ElementNotAttached

Element is not attached to the DOM

ElementNotVisible

Element is not visible

Fields

selector: String

The selector for the element

ElementNotEnabled

Element is not enabled

Fields

selector: String

The selector for the element

ElementNotEditable

Element is not editable

Fields

selector: String

The selector for the element

JsEvaluation(String)

JavaScript evaluation failed

File(Error)

File operation error

InvalidUrl(ParseError)

URL parsing error

DownloadFailed(String)

Resource download failed

ScreenshotFailed(String)

Screenshot operation failed

VideoRecording(String)

Video recording error

Tracing(String)

Tracing error

UnsupportedBrowser(String)

Browser type not supported

BrowserNotFound(String)

Browser not found in installation directory

NotImplemented(String)

Feature not implemented yet

ActionFailed(String)

Action on element failed

Internal(String)

Internal error that shouldnโt normally occur

Implementations

impl Error

pub fn timeout(message: impl Into<String>, timeout_ms: u64) -> Self

Create a timeout error

pub fn timeout_duration(message: impl Into<String>, timeout: Duration) -> Self

Create a timeout error from Duration

pub fn element_not_found(selector: impl Into<String>) -> Self

Create an element not found error

pub fn strict_mode_violation(selector: impl Into<String>, count: usize) -> Self

Create a strict mode violation error

pub fn network(message: impl Into<String>) -> Self

Create a network error

pub fn navigation(message: impl Into<String>) -> Self

Create a navigation error

pub fn invalid_argument(message: impl Into<String>) -> Self

Create an invalid argument error

pub fn not_implemented(feature: impl Into<String>) -> Self

Create a not implemented error

pub fn internal(message: impl Into<String>) -> Self

Create an internal error